chore: extract AlertsList component
